+++ args.py
@@ -1,0 +1,32 @@
+import argparse
+
+def get_args():
+    
+    # Identify required inputs via command line arguments
+    parser = argparse.ArgumentParser(description='Create interlibrary loan transactions in ILLiad for articles in a csv file.')
+    parser.add_argument('email', 
+                        help='The email address of the person who will receive the requested materials. This person must already have a user account in ILLiad.')
+    parser.add_argument('filename', 
+                        help='The name of the file to be read. Must be a .csv file.')
+    parser.add_argument('-p', '--pickup', 
+                        help='The library where the requested materials will be picked up. This is only needed if you are requesting physical materials.',
+                        choices=['Hill', 'Hunt', 'Design', 'Natural Resources', 'Veterinary Medicine', 'Textiles', 'METRC', 'Distance/Extension'])
+    parser.add_argument('-t', '--test', action='store_true',
+                        help='Run the script in test mode to output a report showing which transactions will be created and which will produce errors.')
+    args = parser.parse_args()
+    
+    # Assign command line arguments to variables
+    filename = args.filename
+    email = args.email
+    
+    if args.pickup:
+        pickup = args.pickup
+    else:
+        pickup = ''
+
+    if args.test:
+        test_mode = True
+    else:
+        test_mode = False
+    
+    return email, filename, pickup, test_mode
